added ability to modify main avatar of user

+++ endPoints/deleteImage.php
@@ -9,14 +9,15 @@ $data = json_decode(file_get_contents('php://input'), true);
 $dbName = $data[1]["dbName"];
 $dbNameSelector = $dbName . ".image_url";
 $id = $data[0]["id"];
-$image_url = $Db->selectSqlPrepared("SELECT $dbNameSelector FROM $dbName  WHERE id = $id"); // get image_url drom DB
-$hey = $image_url[0]["image_url"];
-unlink($image_url[0]["image_url"]); // delete image from folder
 
+$image_url = $Db->selectSqlPrepared("SELECT $dbNameSelector FROM $dbName  WHERE id = $id"); // get image_url drom DB
+if(file_exists($image_url[0]["image_url"])){
+unlink($image_url[0]["image_url"]); // delete image from folder
+}
 $insertId = $Db->deleteSql($dbName, $data[0]);
 if($data[1]["dbName"] == "avatars" && $data[1]["is_set"] == 1){ // if it's setted avatar 
   $dataAvatar['avatar_url'] = "../src/img/users/noUser/avatars/avatar.jpg";
-$insertId = $Db->updateSql('users_data', $dataAvatar , "id = " . $data[0]["id_owner"]); // make it default
+ $insertId = $Db->updateSql('users_data', $dataAvatar , "id = " . $data[0]["id_owner"]); // make it default
 }
 $result = [
     'info' => $data,
